+++ FullstackApp-updated/restful-web-services/src/main/java/com/in28minutes/rest/webservices/restfulwebservices/todo/TodoRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface TodoRepository extends JpaRepository<Todo, Integer> {
+public interface TodoRepository extends JpaRepository<Todo, Long> {
     List<Todo> findByAssignedTo(String assignedTo);
 
     List<Todo> findByUsername(String username);
